pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "precision-mlps"
version = "0.1.0"
description = "Machine-precision MLP training via quasi-interpolant theory"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "mpmath>=1.3.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "matplotlib>=3.7"]

[tool.setuptools.packages.find]
include = ["src*", "experiments*"]
